olap_backend_api/requirements/base.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
alabaster==1.0.0
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.9.1
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
billiard==4.2.1
black==25.1.0
boto3==1.40.3
botocore==1.40.3
celery==5.5.3
certifi==2025.8.3
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coverage==7.10.2
cron-descriptor==1.4.5
cryptography==45.0.6
distlib==0.4.0
Django==5.2.5
django-anymail==13.0.1
django-appconf==1.1.0
django-cachalot==2.8.0
django-cache-machine==1.2.0
django-celery-beat==2.8.1
django-compressor==4.5.1
django-cors-headers==4.7.0
django-countries==7.6.1
django-debug-toolbar==6.0.0
django-defender==0.9.8
django-environ==0.12.0
django-extensions==4.1
django-filter==25.1
django-health-check==3.20.0
django-htmlmin==0.11.0
django-livereload-server==0.5.1
django-model-utils==5.0.0
django-phonenumber-field==8.1.0
django-ratelimit==4.1.0
django-redis==6.0.0
django-reversion==5.1.0
django-security==0.9.10
django-silk==5.4.0
django-storages==1.14.6
django-timezone-field==7.1
django_celery_results==2.6.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
django-cors-headers==4.7.0
docutils==0.21.2
dotenv==0.9.9
drf-yasg==1.21.10
et_xmlfile==2.0.0
factory_boy==3.3.3
Faker==37.5.3
filelock==3.18.0
flake8==7.3.0
flower==2.0.1
freezegun==1.5.4
frozenlist==1.7.0
gprof2dot==2025.4.14
gunicorn==23.0.0
h11==0.16.0
html5lib==1.1
humanize==4.12.3
identify==2.6.12
idna==3.10
imagesize==1.4.1
inflection==0.5.1
iniconfig==2.1.0
isort==6.0.1
Jinja2==3.1.6
jmespath==1.0.1
kombu==5.5.4
MarkupSafe==3.0.2
marshmallow==4.0.0
mccabe==0.7.0
multidict==6.6.3
mypy==1.17.1
mypy_extensions==1.1.0
newrelic==10.15.0
nodeenv==1.9.1
numpy==2.3.2
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
pathspec==0.12.1
phonenumbers==9.0.10
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
pre_commit==4.2.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.2
psycopg2-binary==2.9.10
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.1
pytest-cov==6.2.1
pytest-django==4.11.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
rcssmin==1.1.2
redis==6.3.0
reportlab==4.4.3
requests==2.32.4
rjsmin==1.2.2
roman-numerals-py==3.1.0
s3transfer==0.13.1
sentry-sdk==2.34.1
six==1.17.0
snowballstemmer==3.0.1
soupsieve==2.7
Sphinx==8.2.3
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlparse==0.5.3
tornado==6.5.1
twilio==9.7.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
vine==5.1.0
virtualenv==20.33.1
wcwidth==0.2.13
webencodings==0.5.1
whitenoise==6.9.0
yarl==1.20.1